Fix memory corruption bugs
The trace flag was not initialised if there the command-line option is
not set.

More importantly, the stacks and queues in Storage::getTop were passed
by value, such that the memory was already free'd by the time the values
are copied into the atom, leading to all sorts of memory corruption and
nasal demons.